Should fix the problem with 32bits system. This uses smaller integers to make the array fit with a void pointer in any case.

+++ src/game.c
@@ -170,7 +170,7 @@ click(void *data, __UNUSED__ Evas *e, Evas_Object *obj, void *event_info)
 {
    Evas_Event_Mouse_Down *ev = event_info;
    int x, y;
-   int coord[2];
+   uint8_t coord[2];
    char str[16];
 
    /* get back the coordinates of the cell */
